# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: audit — API fiscal (ESAA-Flow)
  version: 1.0.0
  description: |
    Contrato mínimo para destravar o frontend (Lovable). Toda escrita passa pelo
    orquestrador single-writer: o cliente envia uma *intenção*; a API valida
    (7 camadas), grava o evento e re-projeta. Leituras vêm da projeção.

    Convenções:
    - Identificadores em inglês; mensagens de erro em PT-BR.
    - `cnpj` sempre com 14 dígitos, sem máscara.
    - `period` no formato `YYYY-MM` (competência).
    - Toda resposta de escrita devolve o `event_seq` gravado e o `projection_hash`.
servers:
  - url: http://localhost:3000/v1
    description: desenvolvimento
  - url: https://api.SEU-DOMINIO.com.br/v1
    description: produção
security:
  - bearerAuth: []
...
paths:
...
  /clients/{cnpj}/assessments/{period}:
    parameters:
      - $ref: "#/components/parameters/Cnpj"
      - $ref: "#/components/parameters/PeriodPath"
    get:
      tags: [assessment]
      summary: Apuração dual da competência (projeção + hash)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Assessment" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [assessment]
      summary: (Re)projeta a apuração (emite `assessment.projected`)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WriteResult" }
        "409": { $ref: "#/components/responses/Rejected" }
  /clients/{cnpj}/assessments/{period}/adjustments:
    parameters:
      - $ref: "#/components/parameters/Cnpj"
      - $ref: "#/components/parameters/PeriodPath"
    post:
      tags: [assessment]
      summary: Registra ajuste manual justificado (emite `assessment.adjusted`)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Adjustment" }
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WriteResult" }
        "409": { $ref: "#/components/responses/Rejected" }
  /clients/{cnpj}/assessments/{period}/confirm:
    parameters:
      - $ref: "#/components/parameters/Cnpj"
      - $ref: "#/components/parameters/PeriodPath"
    post:
      tags: [assessment]
      summary: Confirma e fecha o período (emite `assessment.confirmed` + `period.closed`; terminal)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projection_hash]
              properties:
                projection_hash:
                  type: string
                  description: Hash que o usuário viu na tela; se divergir do atual, a confirmação é rejeitada (`verification_mismatch`)
                note: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WriteResult" }
        "409": { $ref: "#/components/responses/Rejected" }

  # ------------------------------------------------------ reconciliation
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cnpj:
      in: path
      name: cnpj
      required: true
      # Doze posições alfanuméricas e dois dígitos verificadores numéricos: a
      # Receita emite CNPJ alfanumérico desde 31/07/2026 (IN RFB 2.229/2024).
      # O dígito verificador não é conferido pelo schema — JSON Schema não faz
      # módulo 11 —, e sim na fronteira de entrada da API.
      schema: { type: string, pattern: "^[0-9A-Za-z]{12}[0-9]{2}$" }
...
    PeriodPath:
      in: path
      name: period
      required: true
      schema: { $ref: "#/components/schemas/PeriodId" }
...
  responses:
...
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Rejected:
      description: Intenção rejeitada pelo pipeline (evento `output.rejected` gravado)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Rejection" }
  schemas:
    PeriodId:
      type: string
      pattern: "^[0-9]{4}-(0[1-9]|1[0-2])$"
      example: "2027-01"
    Regime:
      type: string
      enum: [mei, simples_integrado, simples_hibrido, lucro_presumido, lucro_real]
    PeriodState:
      type: string
      enum: [open, assessed, reconciled, confirmed]
...
    RejectionReason:
      type: string
      enum:
        - schema_violation
        - unknown_code
        - code_incompatible
        - regime_violation
        - closed_period_violation
        - fisco_mismatch
        - duplicate_document
        - sequence_gap
        - verification_mismatch
        - boundary_violation
...
    Assessment:
      type: object
      description: |
        Apuração de uma competência, como `GET /clients/{cnpj}/assessments/{period}`
        devolve. Valores em **centavos**, inteiros — nunca reais em ponto
        flutuante, porque um centavo de arredondamento numa apuração é uma
        divergência com o Fisco.
      properties:
        period: { $ref: "#/components/schemas/PeriodId" }
        regime: { $ref: "#/components/schemas/Regime" }
        totals:
          type: object
          description: |
            Um `TaxTotals` por tributo, num mapa plano. As chaves atuais são
            `icms`, `ipi`, `pis`, `cofins` (sistema atual) e `ibs_uf`,
            `ibs_mun`, `cbs` (reforma).
          additionalProperties: { $ref: "#/components/schemas/TaxTotals" }
        total_due_cents:
          type: integer
          nullable: true
          description: |
            Soma dos `dueCents`. **`null` quando qualquer tributo não é
            determinável** — somar tratando o indeterminável como zero daria um
            total que parece calculado e não é.
        not_computable:
          type: array
          description: "O que não foi possível calcular, e por quê. Ver NotComputable."
          items: { $ref: "#/components/schemas/NotComputable" }
        coverage:
          type: object
          description: "Prontidão para a reforma: itens que já trazem o grupo IBS/CBS."
          properties:
            itemsWithReformGroup: { type: integer }
            itemsTotal: { type: integer }
        documents_count: { type: integer }
        items_count: { type: integer }
        projection_hash:
          type: string
          description: |
            Hash da projeção **no momento em que esta apuração foi calculada** —
            não o hash atual do log.
        is_current:
          type: boolean
          description: |
            O `projection_hash` acima ainda é o hash atual do log. Qualquer
            evento posterior (um ajuste, um documento novo) o torna `false`, e
            então `POST .../confirm` com esse hash é recusado com
            `verification_mismatch`, corretamente. A tela deve exigir nova
            projeção e revisão em vez de tentar confirmar.

            A resposta **não** devolve o hash atual, de propósito: entregá-lo
            convidaria a tela a reenviá-lo, que é exatamente o que anularia a
            proteção.
        event_seq: { type: integer }
        computed_at: { type: string, format: date-time }
        adjustments:
          type: array
          items: { $ref: "#/components/schemas/Adjustment" }
    TaxTotals:
      type: object
      description: |
        Quatro números diferentes, e a tela **não deve colapsá-los em um
        "total"**: os dois primeiros saem dos documentos e são certos; os dois
        últimos dependem de norma publicada.
      properties:
        debitsCents:
          type: integer
          description: "Tributo destacado nas saídas. Sempre calculável."
        potentialCreditsCents:
          type: integer
          description: "Tributo destacado nas entradas, antes do juízo de creditamento."
        creditableCents:
          type: integer
          nullable: true
          description: |
            Crédito aproveitável. **`null` quando a regra de creditamento não
            está publicada** para o regime e a competência — ausência de regra
            significa "não sei", não "não credita".
        dueCents:
          type: integer
          nullable: true
          description: |
            `debitsCents - creditableCents`. `null` quando o crédito não é
            determinável. Renderizar como *"não determinável"* com o motivo de
            `not_computable` ao lado: um `—` sem explicação parece defeito, e um
            `0` seria mentira.
        ruleId:
          type: string
          description: "Regra aplicada, para a memória de cálculo."
    NotComputable:
      type: object
      description: |
        Por que um tributo ou item não pôde ser calculado. É o que sustenta o
        `null` em `dueCents` — sem isso, o `null` seria só um vazio.
      properties:
        scope: { type: string, enum: [tributo, item] }
        subject: { type: string, description: "O tributo ou o código do item" }
        reason:
          type: string
          enum: [rule_not_published, missing_reform_group]
          description: |
            `rule_not_published`: nenhuma regra de creditamento vigente para o
            regime e a competência. `missing_reform_group`: o item não traz o
            grupo IBS/CBS, o que é o esperado em documento de 2025.
        message: { type: string, description: "Explicação em PT-BR, pronta para exibir." }
    Adjustment:
      type: object
      required: [tax, amount_cents, reason]
      description: |
        Ajuste manual justificado. **Não sobrescreve a apuração**: entra como
        linha própria, e a diferença entre o apurado e o ajustado é informação.
      properties:
        id: { type: string, readOnly: true }
        tax:
          type: string
          enum: [icms, ipi, pis, cofins, ibs_uf, ibs_mun, cbs]
        amount_cents:
          type: integer
          description: "Positivo aumenta o débito; negativo reduz."
        reason:
          type: string
          minLength: 3
          maxLength: 1000
          description: "Obrigatória. Vai para o event log junto do ajuste."
        reference_access_key: { type: string, nullable: true }
        created_at: { type: string, format: date-time, readOnly: true }
...
    WriteResult:
      type: object
      properties:
        event_id: { type: string, format: uuid }
        event_seq: { type: integer }
        action: { type: string }
        projection_hash: { type: string }
        state: { $ref: "#/components/schemas/PeriodState" }
    Rejection:
      type: object
      properties:
        rejected: { type: boolean, enum: [true] }
        filename:
          type: string
          description: |
            Arquivo que foi recusado. Presente nas rejeições do `207` de
            ingestão — sem ele, num lote de 200 XMLs a tela sabe que houve
            recusa e não sabe de qual arquivo.
        layer: { type: integer, minimum: 1, maximum: 7 }
        reason: { $ref: "#/components/schemas/RejectionReason" }
        message: { type: string }
        details: { type: object }
        rejection_event_seq: { type: integer, description: Seq do `output.rejected` gravado }
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }